requirements.txt
Django==5.1
python-dotenv==1.0.1
django-storages==1.14.4
boto3==1.35.99
